moved enums into their own files (plus EnumCase too)
