+++ lbplanner/classes/enums/NOTIF_TRIGGER.php
@@ -1,5 +1,5 @@
 <?php
-// This file is part of local_lbplanner.
+// This file is part of the local_lbplanner.
 //
 // Moodle is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -14,23 +14,48 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 /**
- * Defines versioning
+ * enum for notif trigger
  *
  * @package local_lbplanner
+ * @subpackage enums
  * @copyright 2024 NecodeIT
  * @license http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
+namespace local_lbplanner\enums;
+
 defined('MOODLE_INTERNAL') || die();
 
-$release = '0.0.0';
+// TODO: revert to native enums once we migrate to php8.
 
-$plugin->component = 'local_lbplanner';
-$plugin->release = 'Alpha v.'.$release;
-$plugin->version = 2024031200;
-$plugin->dependencies = [
-    // Depend upon version 2023110600 of local_modcustomfields.
-    'local_modcustomfields' => 2023110600,
-];
+use local_lbplanner\polyfill\Enum;
 
-set_config('release', $release, 'local_lbplanner');
+/**
+ * Possible triggers for sending a notification
+ */
+class NOTIF_TRIGGER extends Enum {
+    /**
+     * Invitation sent
+     */
+    const INVITE = 0;
+    /**
+     * Invitation accepted
+     */
+    const INVITE_ACCEPTED = 1;
+    /**
+     * Invitation declined
+     */
+    const INVITE_DECLINED = 2;
+    /**
+     * User left the plan
+     */
+    const PLAN_LEFT = 3;
+    /**
+     * User got removed from the plan
+     */
+    const PLAN_REMOVED = 4;
+    /**
+     * User registered
+     */
+    const USER_REGISTERED = 5;
+}
