+++ lbplanner/classes/enums/PLAN_INVITE_STATE.php
@@ -1,5 +1,5 @@
 <?php
-// This file is part of local_lbplanner.
+// This file is part of the local_lbplanner.
 //
 // Moodle is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -14,23 +14,40 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 /**
- * Defines versioning
+ * enum for plan invite state
  *
  * @package local_lbplanner
+ * @subpackage enums
  * @copyright 2024 NecodeIT
  * @license http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
+namespace local_lbplanner\enums;
+
 defined('MOODLE_INTERNAL') || die();
 
-$release = '0.0.0';
+// TODO: revert to native enums once we migrate to php8.
 
-$plugin->component = 'local_lbplanner';
-$plugin->release = 'Alpha v.'.$release;
-$plugin->version = 2024031200;
-$plugin->dependencies = [
-    // Depend upon version 2023110600 of local_modcustomfields.
-    'local_modcustomfields' => 2023110600,
-];
+use local_lbplanner\polyfill\Enum;
 
-set_config('release', $release, 'local_lbplanner');
+/**
+ * States an invite can be in
+ */
+class PLAN_INVITE_STATE extends Enum {
+    /**
+     * pending invite
+     */
+    const PENDING = 0;
+    /**
+     * accepted invite
+     */
+    const ACCEPTED = 1;
+    /**
+     * declined invite
+     */
+    const DECLINED = 2;
+    /**
+     * expired invite
+     */
+    const EXPIRED = 3;
+}
