+++ lbplanner/classes/polyfill/Enum.php
@@ -28,11 +28,12 @@ defined('MOODLE_INTERNAL') || die();
 
 use ReflectionClass;
 use ValueError;
+use lb_planner_local\polyfill\EnumCase;
 
 /**
  * Class which is meant to serve as a substitute for native enums.
  */
-abstract class Enum {
+class Enum {
     /**
      * tries to match the passed value to one of the enum values
      * @param mixed $value the value to be matched
@@ -87,25 +88,5 @@ abstract class Enum {
         }
         $result[-1] = ']';
         return $result;
-    }
-}
-
-/**
- * This represents a single case within an Enum
- */
-class EnumCase {
-    /** @var string the name of the case */
-    public string $name;
-    /** @var string the value of the case */
-    public mixed $value;
-    /**
-     * Constructs an EnumCase
-     *
-     * @param string $name the name of the case
-     * @param mixed $value the value of the case
-     */
-    public function __construct(string $name, mixed $value) {
-        $this->name = $name;
-        $this->value = $value;
     }
 }
