+++ lbplanner/services/plan/accept_invite.php
@@ -21,9 +21,7 @@ use external_function_parameters;
 use external_value;
 use local_lbplanner\helpers\plan_helper;
 use local_lbplanner\helpers\notifications_helper;
-use local_lbplanner\helpers\PLAN_ACCESS_TYPE;
-use local_lbplanner\helpers\PLAN_INVITE_STATE;
-use local_lbplanner\helpers\NOTIF_TRIGGER;
+use local_lbplanner\enums\{PLAN_ACCESS_TYPE,PLAN_INVITE_STATE,NOTIF_TRIGGER};
 
 /**
  * Accept an invite to the plan.
