+++ lbplanner/services/plan/decline_invite.php
@@ -19,10 +19,9 @@ namespace local_lbplanner_services;
 use external_api;
 use external_function_parameters;
 use external_value;
-use local_lbplanner\helpers\NOTIF_TRIGGER;
 use local_lbplanner\helpers\plan_helper;
 use local_lbplanner\helpers\notifications_helper;
-use local_lbplanner\helpers\PLAN_INVITE_STATE;
+use local_lbplanner\enums\{NOTIF_TRIGGER,PLAN_INVITE_STATE};
 
 /**
  * Decline an invite from the plan.
