+++ lbplanner/services/user/delete_user.php
@@ -19,12 +19,9 @@ namespace local_lbplanner_services;
 use dml_exception;
 use external_api;
 use external_function_parameters;
-use external_single_structure;
 use external_value;
-use local_lbplanner\helpers\user_helper;
-use local_lbplanner\helpers\plan_helper;
-use local_lbplanner\helpers\course_helper;
-use local_lbplanner\helpers\notifications_helper;
+use local_lbplanner\helpers\{user_helper,plan_helper,course_helper,notifications_helper};
+use local_lbplanner\enums\{PLAN_INVITE_STATE,PLAN_ACCESS_TYPE};
 use moodle_exception;
 
 /**
@@ -81,7 +78,7 @@ class user_delete_user extends external_api {
         if (
             !(count(plan_helper::get_plan_members($planid)) == 1 )
             &&
-            !(plan_helper::get_access_type($planid, $userid) == plan_helper::ACCESS_TYPE_OWNER)) {
+            !(plan_helper::get_access_type($planid, $userid) == PLAN_ACCESS_TYPE::OWNER)) {
             self::call_external_function('local_lbplanner_plan_leave_plan', ['userid' => $userid, 'planid' => $planid]);
         }
         $DB->delete_records(plan_helper::DEADLINES_TABLE, ['planid' => $planid]);
