+++ lbplanner/services/user/register_user.php
@@ -22,12 +22,8 @@ use external_function_parameters;
 use external_single_structure;
 use external_value;
 use invalid_parameter_exception;
-use local_lbplanner\helpers\NOTIF_TRIGGER;
-use local_lbplanner\helpers\PLAN_ACCESS_TYPE;
-use local_lbplanner\helpers\PLAN_EK;
-use local_lbplanner\helpers\user_helper;
-use local_lbplanner\helpers\plan_helper;
-use local_lbplanner\helpers\notifications_helper;
+use local_lbplanner\helpers\{user_helper,plan_helper,notifications_helper};
+use local_lbplanner\enums\{NOTIF_TRIGGER,PLAN_ACCESS_TYPE,PLAN_EK};
 use moodle_exception;
 use stdClass;
